Updated the interface to work with recent versions of openmpi.
The interface should now work with openmpi versions 1.8 through to 4.x,
however, it has only been tested on openmpi versions 2.1.1, 4.0.1,
and 4.0.2.

The changes are mainly to simplify the interface by only supporting
openmpi 1.8 and greater and always providing the MPI_FILE_INTERFACE.
This eliminated much of the logic, some of which seemed wrong for
openmpi versions greater than 1.x.